Moved constants file.
