+++ src/main/webapp/creatives/simid_protocol.js
@@ -5,7 +5,7 @@
  * Note: File originally from https://github.com/InteractiveAdvertisingBureau/SIMID/tree/master/examples
  */
 
-import {PlayerMessage, CreativeMessage} from '../creatives/constants.js';
+import {PlayerMessage, CreativeMessage} from './constants.js';
 
 export default class SimidProtocol {
